Update blech_clust.py
line 246: after running blech_exp_info.py with no laser DIN, I was getting the value of laser_dig_in as [], an empty array, not None.
I'm not sure if this is universal, so I modified 246 to ignore laser_dig_in if it is EITHER None or [].

line 251: running blech_exp_info.py, I had a lot more DINs than I actually used, so most of them were ignored. This creates an issue when plotting the DINs; all of the DINs are plotted whether or not they're selected to be used (I think this behavior is good as a diagnostic tool). yticks are only plotted for selected DINs (again, seems good), but when naming the yticks, np.arange(len(dig_in_map)) just returns a 0:N array regardless of the indices of the selected DINs, so the yticks may be labeling the wrong DINs if the selected DINs aren't 0:N. So, changing that to an array of the key  positions in dig_in_map fixes that behavior.