Update blech_process.py
Tidied up a redundant use of `import imp_metadata`

Added a thing that suppresses the pickle version warning
(may be worth revisiting? would it be better to update the old pickle, or impose an older version of pickle in the requirements of blech_clust? Anyway, this works in the meantime)