Update drift_check.py
Major Change:
Adapted spike-train plotting to work with sessions with uneven #s of trials per stimulus; finds the stimulus with the most trials, and then pads out the shorter spike trains with NAs. The padded object, unit_spike_trains, and objects created from it, are only used in plotting, as far as I can tell, so adding the NAs shouldn't cause issues further down in the script.

Minor changes:
Added some comments describing the shape of different data objects, for my reference
Changed the sizing of the spike trains plot to be responsive to # of units to be plotted: for large sets of units, expands the plot to fit them.