Account for flexible split inputs from CSV
Pre-commit behavior first tests whether there's any value in the split column, initiates splitting, and produces an error if the column value does not contain a numeral

Altered the first test to ignore whitespace inputs; "   " previously triggered splitting, and does not now.
Additionally, once splitting has been initiated, checks if there is a numeral in the csv input. If there is a numeral, it asks for user input but defaults to that numeral if none is given. If there is no numeral, it asks for user input and defaults to 5, as with manual sorting.

This is now pretty airtight, I think.