Clean up the test for EMG data
`raw_emg_electrodes` is pulled directly from the `raw_emg` group of the hdf5 file earlier in the script, and testing it's length is actually also used earlier to check the presence or absence of EMG data. That seems elegant, so I re-used it here.

This also takes into account the fact that `raw_emg` is added to the hdf5 file by default in `blech_clust.py`, so it's always there, and sometimes empty, but checking if it exists isn't useful.