Fix checks for laser
As-is, `laser_durations` is a list of length {trials}, and if there's no laser, the list is full of 0s. This still gives it a non-zero length, but the sum is 0. Presumably for sessions WITH lasers, the duration is > 0, so the sum of the trials would be greater than 0?

This has been tested with a dataset without lasers, and worked, but I haven't tried it with laser data.