blech_post_process_utils.py: drop `level0` from sort file, and make split more permissive
A column `level0` is added to the cell sorting csv after running, and if post-process is re-run, this causes an error. So this deletes that column if it exists before trying to create it.

Also, made the split info more permissive; shouldn't throw an error if split is anything other than numeric.